api erro in order

+++ app/api/order/[id]/route.ts
@@ -1,14 +1,20 @@
-import {  NextResponse } from "next/server";
+"@ts-expect-error"
+import { getUserFromToken } from "@/lib/auth";
 import { connectDB } from "@/lib/mongodb";
 import Order from "@/models/Order";
-
+import {   NextRequest, NextResponse } from "next/server";
 export async function DELETE(
-  context: { params: { id: string } }
+  req: NextRequest,
+  { params }: { params: { id: string } }
 ) {
   try {
     await connectDB();
+    const user = await getUserFromToken(req);
+    if(!user){
+      return NextResponse.json({ message: "user not found" }, { status: 404 });
 
-    const { id } = context.params;
+    }
+    const { id } = params;
     const deletedOrder = await Order.findByIdAndDelete(id);
 
     if (!deletedOrder) {
@@ -16,10 +22,10 @@ export async function DELETE(
     }
 
     return NextResponse.json({ message: "Order deleted", order: deletedOrder }, { status: 200 });
-  } catch (error: unknown) {
-    if (error instanceof Error) {
-      return NextResponse.json({ error: error.message }, { status: 500 });
-    }
-    return NextResponse.json({ error: "Something went wrong" }, { status: 500 });
+  } catch (error) {
+    console.error("Error deleting order:", error);
+    return NextResponse.json({ message: "Internal server error" }, { status: 500 });
   }
 }
+
+
